+++ src/js/FiniteStateMachine/RectSelect.js
@@ -107,9 +107,10 @@ export class RectSelect extends FSM {
 
             // execute general callback
             if (view.aladin.callbacksByEventName) {
-                var callback = view.aladin.callbacksByEventName['select'];
+                var callback = view.aladin.callbacksByEventName['objectsSelected'] || view.aladin.callbacksByEventName['select'];
                 if (typeof callback === "function") {
                     let objList = Selector.getObjects(s, view);
+                    view.selectObjects(objList);
                     callback(objList);
                 }
             }