+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelogs
 
-## Unreleased
+## 3.4.5-beta
+
 * [feat] add `layerChanged` event when a layer is added or removed
+* [deprecate] of `select` event, use `objectsSelected` event instead
+* [ui] add the ability to switch the tile format to download 
 
 ## 3.4.3-beta